refactor: User객체, service, JCFUserService 이동 및 구조 변경
